Handle all zero data case.
If all of the data is zero for the Sparkline, the "y range" is
calculated to 0. The range is used as the denominator when calculating
the scaling factor, so it obviously fails.
The Line and Area graphs use the code in SVG::Sparkline::Utils to do
this calculation. The Bar graph does it directly.